Fix dist computation.

+++ examples/rl/platoon/inference/contrib_policy/filter_obs.py
@@ -151,7 +151,7 @@ def wps_to_pixels(
     wps_overlay = np.array([w / 2, h / 2, 0]) + wps_pixels * np.array([1, -1, 1])
     wps_rfloat = np.rint(wps_overlay)
     wps_valid = wps_rfloat[(wps_rfloat[:,0] >= 0) & (wps_rfloat[:,0] < w) & (wps_rfloat[:,1] >= 0) & (wps_rfloat[:,1] < h)] 
-    wps_rint = wps_valid.astype(np.uint8)
+    wps_rint = wps_valid.astype(int)
     return wps_rint
     # fmt: on
